[RTL][RISCV] Merge load/store insns
We're interested in grouping SET insns that access memory and have
exactly one memory reference. First, we pair insns. Two paired insns are
either load or store SET insns that access the adjacent memory
locations. One insn can belong to two pairs, as long as it represents
the "left" member in the first pair and the "right" member in the second
pair. If that's the case, two pairs can form one group of three.
Grouping continues until we examine all group combinations.

The transformation phase begins by scheduling insns based on their order
in the basic block. Each insn is scheduled as soon as all statements on
which it is dependent have been scheduled. For loadstore groups, this
property must be satisfied for each insn in the group. Whenever a group
of scalar insns is scheduled, a new SIMD operation is emitted instead.